requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.7
coloredlogs==15.0.1
crashtest==0.4.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
dulwich==0.21.7
durationpy==0.9
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
graphlib_backport==1.1.0
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
installer==0.7.0
ipython==8.12.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==24.3.1
kubernetes==31.0.0
langchain==0.2.17
langchain-chroma==0.1.4
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.1
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.57.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkginfo==1.12.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
posthog==3.7.4
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.4.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trove-classifiers==2024.10.21.16
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.0
xattr==1.1.0
yarg==0.1.9
yarl==1.15.2
zipp==3.20.2
